+++ src/test/java/dk/kb/license/integrationtest/RightsModuleIntegrationTest.java
@@ -125,8 +125,9 @@ public class RightsModuleIntegrationTest extends DsLicenseUnitTestUtil {
         restrictedId.setIdValue("ds.tv:oai:io:e027e1dc-5006-4f54-b2b7-ec451940c500");
         restrictedId.setIdType(IdTypeEnumDto.DS_ID);
         restrictedId.setPlatform(PlatformEnumDto.DRARKIV);
-        RightsModuleFacade.createRestrictedId(restrictedId, false);
+        restrictedId.setComment("Brugeren har trukket deres samtykke tilbage");
 
+        RightsModuleFacade.createRestrictedId(restrictedId, false);
         RestrictedIdOutputDto outputObject = RightsModuleFacade.getRestrictedId(restrictedId.getIdValue(), restrictedId.getIdType(), restrictedId.getPlatform());
 
         int deleted = RightsModuleFacade.deleteRestrictedId(outputObject.getId().longValue(), false);
@@ -140,6 +141,7 @@ public class RightsModuleIntegrationTest extends DsLicenseUnitTestUtil {
         restrictedId.setIdValue("00123466486");
         restrictedId.setIdType(IdTypeEnumDto.DR_PRODUCTION_ID);
         restrictedId.setPlatform(PlatformEnumDto.DRARKIV);
+        restrictedId.setComment("Brugeren har trukket deres samtykke tilbage");
 
         RightsModuleFacade.createRestrictedId(restrictedId, false);
         RestrictedIdOutputDto outputRight = RightsModuleFacade.getRestrictedId(restrictedId.getIdValue(), IdTypeEnumDto.DR_PRODUCTION_ID, PlatformEnumDto.DRARKIV);
@@ -156,6 +158,7 @@ public class RightsModuleIntegrationTest extends DsLicenseUnitTestUtil {
         restrictedId.setIdValue("1234664800");
         restrictedId.setIdType(IdTypeEnumDto.DR_PRODUCTION_ID);
         restrictedId.setPlatform(PlatformEnumDto.DRARKIV);
+        restrictedId.setComment("Brugeren har trukket deres samtykke tilbage");
 
         RightsModuleFacade.createRestrictedId(restrictedId, false);
         RestrictedIdOutputDto outputRight = RightsModuleFacade.getRestrictedId(restrictedId.getIdValue(), IdTypeEnumDto.DR_PRODUCTION_ID, PlatformEnumDto.DRARKIV);
@@ -172,6 +175,7 @@ public class RightsModuleIntegrationTest extends DsLicenseUnitTestUtil {
         restrictedId.setIdValue("123466489");
         restrictedId.setIdType(IdTypeEnumDto.DR_PRODUCTION_ID);
         restrictedId.setPlatform(PlatformEnumDto.DRARKIV);
+        restrictedId.setComment("Brugeren har trukket deres samtykke tilbage");
 
         RightsModuleFacade.createRestrictedId(restrictedId, false);
         RestrictedIdOutputDto outputRight = RightsModuleFacade.getRestrictedId(restrictedId.getIdValue(), IdTypeEnumDto.DR_PRODUCTION_ID, PlatformEnumDto.DRARKIV);